Introduce --no-grub-install option
This option is useful when the grub.cfg file on the EFI partition
needs to be recreated without re-installing grub. This happens in
kiwi, which currently works around the issue by moving grub-install
out of the way. When setting up full disk encryption with a TPM,
we need to recreate grub.cfg as well - but we may not be able to
move grub-install if the root FS is readonly. Which happens with
transactional systems.

Hence this option.

Signed-off-by: Olaf Kirch <okir@suse.com>