Use the long name to specify the grub2 key protector
Grub2 upstream merged a series of patches to provide cryptomount the key
file to unlock the disk. Unfortunately, the short name (-k) of "--keyfile"
conflicts with the one chosen by the grub2 TPM2 disk unlock patches. To
avoid the compatibility issues, always use the long name (--protector)
to specify the key protector for cryptomount.

Signed-off-by: Gary Lin <glin@suse.com>