Always use the removable way for SL-Micro
Since SL-Micro is image-based, it's designed to boot from the default
boot path i.e. \EFI\BOOT\boot<arch>.efi. To make the behavior consistent
and avoid breaking the TPM sealed key, this commit turns "removable" to
"yes" when the system is "SL-Micro" and skips the creation of the boot
option. Besides, it reuses the boot option deletion code to make sure
that no 'sl-secureboot' boot option exists.

FIXES: bsc#1230316

Signed-off-by: Gary Lin <glin@suse.com>