Limit the removable option to encrypted SL-Micro
For the normal SL-Micro images, it doesn't matter whether the default
boot path (\EFI\BOOT\boot<arch>.efi) is used or not as long as the
firmware can find the correct bootloaders.

This commit adds the addition check for 'GRUB_TPM2_SEALED_KEY' to make
sure that TPM2 unsealing is used in the system and marks such system as
'removable'.

Signed-off-by: Gary Lin <glin@suse.com>